init golang modules from existing 'go get' calls
to recreate:
 - `git clone` a fresh checkout, not in gopath
 - `make godeps` to download all deps as they currently exist
 - `go mod init github.com/hyperledger-labs/fabric-private-chaincode`
   in base directory.

`go mod download` can be used to initialize the go build cache in a
similar manner to the existing `go get` calls.

As it existed the `go get` commands in the makefile have the potential
to float and not get any specific version. The newer go toolchains
that are module enabled support giving version specifiers to `go get`,
for example `go get github.com/spf13/viper@v1.6.2`. Another behavior
of module enabled toolchain, is that when in a module, `go get` will
write a dependency to the `go.mod` file. Combining these two facts, if
we enable modules, and then run the existing `go get` commands, we
will end up with the dependencies currently necessary to build. After
which, we use `go mod download` to download all modules in the
`go.mod` file. This change replaces all `go get` invocations with a
single invocation of `go mod download`.

There are probably actual used dependencies (retrieved when compiling
the code) that are still missing.

resolves hyperledger-labs/fabric-private-chaincode#94

Signed-off-by: Morgan Bauer <mbauer@us.ibm.com>