Move ecc init logic into setup
With PR #330 we removed support for chaincode init; however, init in ecc
write the fpc chaincode MRENCLAVE to the ledger that is used during
attestation validtion inside ERCC and TLCC; PR #330 missed to move the
logic so this commit fixes it. Note that in the future that will change
and MRENCLAVE will be fetched from chaincode definition.

Signed-off-by: bur <bur@zurich.ibm.com>