Fix configuration path for case-sensitive fs (#660)
* Fix configuration path for case-sensitive fs

The client SDK needs to load the configuration file from the file
system. In this sample app we use the configurations provided in
fabric-samples/test-network. On a case-sensitive fs (Linux), the
constructed path (ccpPath) in this demo app is not correct. We fix
this by demo by using strings.ToLower as we know the path already.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>